PXB-3185 - MySQL 8.2 Merge

+++ sql/sql_const_folding.cc
@@ -41,12 +41,12 @@
 #include "decimal.h"                       // E_DEC_FATAL_ERROR
 #include "field_types.h"                   // MYSQL_TYPE_DATE
                                            // assert
-#include "my_decimal.h"                    // my_decimal, my_decimal_cmp
 #include "my_inttypes.h"                   // longlong, ulonglong
 #include "my_time.h"                       // TIME_to_longlong_datetime_packed
 #include "mysql/strings/dtoa.h"            // DECIMAL_NOT_SPECIFIED
 #include "mysql/udf_registration_types.h"  // INT_RESULT, STRING_RESULT
 #include "mysql_time.h"                    // MYSQL_TIME
+#include "sql-common/my_decimal.h"         // my_decimal, my_decimal_cmp
 #include "sql/field.h"                     // Field_real, Field
 #include "sql/item.h"                      // Item, Item_field, Item_int
 #include "sql/item_cmpfunc.h"              // Item_bool_func2, Item_cond
@@ -724,11 +724,12 @@ static bool analyze_year_field_constant(THD *thd, Item **const_val,
     case REAL_RESULT:
     case INT_RESULT: {
       const double year = (*const_val)->val_real();
-      if (year > Field_year::MAX_YEAR) {
+      if (year > static_cast<double>(Field_year::MAX_YEAR)) {
         *place = RP_OUTSIDE_HIGH;
       } else if (year < 0.0) {
         *place = RP_OUTSIDE_LOW;
-      } else if (year > 0.0 && year < Field_year::MIN_YEAR) {
+      } else if (year > 0.0 &&
+                 year < static_cast<double>(Field_year::MIN_YEAR)) {
         /*
           These values can be given as constants, but are not allowed to be
           stored in the field, so an = or <> comparison can be folded. For
